fiddle with data types
there's not much reason to do this other than "why not"··· it's not like it's gonna optimize the final executable, in fact it somehow is even larger now that i made that change, but i guess this really makes sure all variables are the exact size i want regardless of implementation, and i guess it also saves on a few keystrokes

+++ shoot.h
@@ -21,7 +21,7 @@ typedef enum trans_t {
 extern int key;
 
 // level tracking
-extern unsigned int levels;
+extern uint8_t levels;
 extern uintptr_t level;
 
 // general engine functions
